Move features away from libs dir
