+++ src/content.js
@@ -4,22 +4,22 @@ import {applyToLink as shortenLink} from 'shorten-repo-url';
 import select from 'select-dom';
 import domLoaded from 'dom-loaded';
 
-import markUnread from './libs/mark-unread';
-import addOpenAllNotificationsButton from './libs/open-all-notifications';
+import markUnread from './features/mark-unread';
+import addOpenAllNotificationsButton from './features/open-all-notifications';
 import addUploadBtn from './libs/upload-button';
-import enableCopyOnY from './libs/copy-on-y';
-import addReactionParticipants from './libs/reactions-avatars';
-import showRealNames from './libs/show-names';
-import addCopyFilePathToPRs from './libs/copy-file-path';
-import addFileCopyButton from './libs/copy-file';
-// - import copyMarkdown from './libs/copy-markdown';
-import linkifyCode from './libs/linkify-urls-in-code';
-import autoLoadMoreNews from './libs/auto-load-more-news';
-import addOPLabels from './libs/op-labels';
-import addMoreDropdown from './libs/more-dropdown';
-import addReleasesTab from './libs/add-releases-tab';
-import scrollToTopOnCollapse from './libs/scroll-to-top-on-collapse';
-import removeDiffSigns from './libs/remove-diff-signs';
+import enableCopyOnY from './features/copy-on-y';
+import addReactionParticipants from './features/reactions-avatars';
+import showRealNames from './features/show-names';
+import addCopyFilePathToPRs from './features/copy-file-path';
+import addFileCopyButton from './features/copy-file';
+// - import copyMarkdown from './features/copy-markdown';
+import linkifyCode from './features/linkify-urls-in-code';
+import autoLoadMoreNews from './features/auto-load-more-news';
+import addOPLabels from './features/op-labels';
+import addMoreDropdown from './features/more-dropdown';
+import addReleasesTab from './features/add-releases-tab';
+import scrollToTopOnCollapse from './features/scroll-to-top-on-collapse';
+import removeDiffSigns from './features/remove-diff-signs';
 import * as linkifyBranchRefs from './features/linkify-branch-refs';
 import hideEmptyMeta from './features/hide-empty-meta';
 import hideOwnStars from './features/hide-own-stars';
