+++ src/features/scroll-to-top-on-collapse.js
